Conolidate leaf validation under Token (#176)
`Token` is a natural point to organize all leaf validation logic.
The validation are all intented with a type in mind
except for the case where we blanket validate for everything.
So the methods for each type should be throwing
and caller can decide whether to ignore the thrown error
and move on to the next candidate.

Token being the funnel point also improves error messages.

The only downside is we are passing more than 2 String indices
around now.
So that might be slower.