Split the offset date-time strategies
... into a strategy for TimeInterval, and a strategy for Date.

This is because the preferences for decoding the 2 are agnostic.
It's valid to want some offset date-time to be interperted as time
interval/double, and some as dates. And both are also agnostic to
using `TOMLDecoder.OffsetDateTime` directely.

Give the strategies a shallower parent type, and include the word
"Strategy" as suffix in their names. This avoids confusion between,
for example, `Fonudation.TimeInterval`, and
`TOMLDecoder.Strategy.TimeInterval`, when the module/namespace is
elided.

Also add docs.